backend/requirements.txt
# XR2Text - Requirements
# Install with: pip install -r requirements.txt

# ============================================
# Core Deep Learning (CUDA 12.x for RTX 4060)
# ============================================
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# ============================================
# Transformers & NLP
# ============================================
transformers>=4.36.0
tokenizers>=0.15.0
datasets>=2.16.0
huggingface-hub>=0.20.0

# ============================================
# Computer Vision
# ============================================
timm>=0.9.12
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1

# ============================================
# API Framework
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================
# Data Processing
# ============================================
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# ============================================
# Training & Monitoring
# ============================================
tensorboard>=2.15.0
wandb>=0.16.0
tqdm>=4.66.0
accelerate>=0.25.0

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.0
pyyaml>=6.0.1
loguru>=0.7.2
requests>=2.31.0

# ============================================
# Evaluation Metrics
# ============================================
nltk>=3.8.1
rouge-score>=0.1.2
sacrebleu>=2.4.0

# ============================================
# Visualization & Notebooks (Research Paper)
# ============================================
matplotlib>=3.8.0
seaborn>=0.13.0
wordcloud>=1.9.0
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0

# ============================================
# Testing
# ============================================
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
